backend/requirements.txt
supabase
fastapi
uvicorn
pydantic
python-dotenv
requests
pydantic[email]